Added NETWORK LOG for Vimeo.com manage video URL. If not recording the screen, I get Technical Difficulties message
I am sharing this log to provide it to Vimeo.com support. Maybe they can pin point which files are being injected by the hackers. I am a target of persistent hacking, and I live with a middle man in every device I use. They disable me from operating websites as they should work, they also have resources with people who are around me, even if I move to another town. I believe they also have resources with the local police departments through the country, as well as many members of any given community. You have to be a target to see it, and live it for a long time (for me this has been happening for over 7 years) to understand it.